refactor endpoints, pase las validaciones al servic, para separar la logica

+++ Application/Services/CivilizationService.cs
@@ -1,25 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Application.Interfaces;
+﻿using Application.Interfaces;
 using Application.Models.Dto;
 using Application.Models.Request;
 using Domain.Entities;
-using Domain.Enums;
 using Domain.Interfaces;
-using Infrastructure.Repositories;
-using Microsoft.EntityFrameworkCore;
-using static Application.Models.Dto.BattleTableDto;
-
-
 
 namespace Application.Services
 {
     public class CivilizationService : ICivilizationService
     {
         private readonly ICivilizationRepository _civilizationRepository;
+
         public CivilizationService(ICivilizationRepository civilizationRepository)
         {
             _civilizationRepository = civilizationRepository;
@@ -38,21 +28,35 @@ namespace Application.Services
         }
 
         public async Task<Civilization> CreateCivilization(CreateCivilizationRequest request, CancellationToken ct)
+
         {
             var civilization = CreateCivilizationRequest.ToEntity(request);
             return await _civilizationRepository.CreateCivilization(civilization, ct);
         }
 
-
-        public async Task UpdateCivilization(int id, UpdateCivilizationRequest request, CancellationToken ct)
+        public async Task<bool> UpdateCivilization(int id, UpdateCivilizationRequest request, CancellationToken ct)
         {
             var civilization = await _civilizationRepository.GetCivizlizationById(id, ct);
-             await _civilizationRepository.UpdateCivilization(civilization, ct);
+            if (civilization is null)
+            {
+                return false;
+            }
+            UpdateCivilizationRequest.ApplyToEntity(request, civilization);
+            await _civilizationRepository.UpdateCivilization(civilization, ct);
+            return true;
         }
-        public async Task DeleteCivilization(int id, CancellationToken ct)
+
+        public async Task<bool> DeleteCivilization(int id, CancellationToken ct)
         {
             var civilization = await _civilizationRepository.GetCivizlizationById(id, ct);
-             await _civilizationRepository.DeleteCivilization(civilization, ct);
+            if (civilization is null)
+            {
+                return false;
+            }
+
+            await _civilizationRepository.DeleteCivilization(civilization, ct);
+
+            return true;
         }
     }
 }
